reorganize imports and traits
