+++ src/create_index/create_index.rs
@@ -1,7 +1,9 @@
 use crate::{
-  behavior::{push_unique, Concat, TransactionQuery},
+  behavior::TransactionQuery,
+  concat::Concat,
   fmt,
   structure::{CreateIndex, CreateIndexParams, LogicalOperator},
+  utils::push_unique,
 };
 
 impl TransactionQuery for CreateIndex {}
