+++ src/delete/delete_internal.rs
@@ -1,33 +1,10 @@
-#[cfg(any(feature = "postgresql", feature = "sqlite"))]
-use crate::behavior::ConcatCommon;
 use crate::{
-  behavior::{concat_raw_before_after, Concat, ConcatSqlStandard},
+  concat::{concat_raw_before_after, sql_standard::ConcatWhere, Concat},
   fmt,
   structure::{Delete, DeleteClause},
 };
 
-impl ConcatSqlStandard<DeleteClause> for Delete {}
-
-impl Delete {
-  fn concat_delete_from(&self, query: String, fmts: &fmt::Formatter) -> String {
-    let fmt::Formatter { lb, space, .. } = fmts;
-    let sql = if self._delete_from.is_empty() == false {
-      let table_name = &self._delete_from;
-      format!("DELETE FROM{space}{table_name}{space}{lb}")
-    } else {
-      "".to_string()
-    };
-
-    concat_raw_before_after(
-      &self._raw_before,
-      &self._raw_after,
-      query,
-      fmts,
-      DeleteClause::DeleteFrom,
-      sql,
-    )
-  }
-}
+impl ConcatWhere<DeleteClause> for Delete {}
 
 impl Concat for Delete {
   fn concat(&self, fmts: &fmt::Formatter) -> String {
@@ -70,5 +47,31 @@ impl Concat for Delete {
   }
 }
 
+impl Delete {
+  fn concat_delete_from(&self, query: String, fmts: &fmt::Formatter) -> String {
+    let fmt::Formatter { lb, space, .. } = fmts;
+    let sql = if self._delete_from.is_empty() == false {
+      let table_name = &self._delete_from;
+      format!("DELETE FROM{space}{table_name}{space}{lb}")
+    } else {
+      "".to_string()
+    };
+
+    concat_raw_before_after(
+      &self._raw_before,
+      &self._raw_after,
+      query,
+      fmts,
+      DeleteClause::DeleteFrom,
+      sql,
+    )
+  }
+}
+
 #[cfg(any(feature = "postgresql", feature = "sqlite"))]
-impl ConcatCommon<DeleteClause> for Delete {}
+use crate::concat::non_standard::{ConcatReturning, ConcatWith};
+
+#[cfg(any(feature = "postgresql", feature = "sqlite"))]
+impl ConcatReturning<DeleteClause> for Delete {}
+#[cfg(any(feature = "postgresql", feature = "sqlite"))]
+impl ConcatWith<DeleteClause> for Delete {}
