+++ src/drop_table/drop_table.rs
@@ -1,7 +1,9 @@
 use crate::{
-  behavior::{push_unique, Concat, TransactionQuery},
+  behavior::TransactionQuery,
+  concat::Concat,
   fmt,
   structure::{DropTable, DropTableParams},
+  utils::push_unique,
 };
 
 impl TransactionQuery for DropTable {}
