+++ src/drop_table/drop_table_internal.rs
@@ -1,10 +1,19 @@
 use crate::{
-  behavior::{concat_raw_before_after, Concat, ConcatSqlStandard},
+  concat::{concat_raw_before_after, Concat},
   fmt,
   structure::{DropTable, DropTableParams},
 };
 
-impl ConcatSqlStandard<DropTableParams> for DropTable {}
+impl Concat for DropTable {
+  fn concat(&self, fmts: &fmt::Formatter) -> String {
+    let mut query = "".to_string();
+
+    query = self.concat_raw(query, &fmts, &self._raw);
+    query = self.concat_drop_table(query, &fmts);
+
+    query.trim_end().to_string()
+  }
+}
 
 impl DropTable {
   fn concat_drop_table(&self, query: String, fmts: &fmt::Formatter) -> String {
@@ -42,16 +51,5 @@ impl DropTable {
       DropTableParams::DropTable,
       sql,
     )
-  }
-}
-
-impl Concat for DropTable {
-  fn concat(&self, fmts: &fmt::Formatter) -> String {
-    let mut query = "".to_string();
-
-    query = self.concat_raw(query, &fmts, &self._raw);
-    query = self.concat_drop_table(query, &fmts);
-
-    query.trim_end().to_string()
   }
 }
